Simplify API by removing focussing on the resource itself
The resource is the missing piece of the web programming puzzle. For that
reason, it's better to retain Ring's API as far as possible and allow routing to
occur outside of Spin. Once a target resource has located, and its corresponding
map established, then Spin's request handler can take over.

This does not preclude users from using routers to locate resources. A database
lookup of a resource now just becomes another router. Spin is agnostic.