Add experiment to research the 'library-over-framework' approach
This is an experimental namespace, currently called spinner but will be renamed
and promoted if successful.

The idea, from @SevereOverfl0w, is to write spin in terms of a HOWTO, possibly
driven by resource and representation data, without callback functions.

This makes spin do less, but preserving user choice. Spin would still exist in
the form of convenience libraries that, at most, transform data structures.